fix typo that prevented this method being called
The super-class method, which we are overriding here, adds
a 'filter' key:value to the 'optical_elements' dictionary of
the config. 1M0-NRES-SCICAM instruments are defined in ConfigDB
to not have any optical_elements. So, having a 'filter' key:value
invalidates the request in the ObservationPortal. Our overriden
method here removes the 'optical_element' dictionary from the request
which is appropriate for 1M-NRES_SCICAM instruments.