Allow cadence to be initialized with non-existant observation_id
rundcadencestrategies gets the latest observation for a cadence to
use it's values for the next observation of the cadence. It calls
the facility to get the latest observation status so it know how to
proceed with the cadence. The current method for starting Photometric
Standard Cadences involves "seeding" the cadence with an
ObservationRecord (with a fake (-1) observation_id). This change
allows the runcadencestrategies to be continue with such
ObservationRecords. (They won't be found in at the Facilty, but
that ok in this case). In the case, that the
ObservationRecord.observation_id is valid but can't be updated from
the facility, the most recent ObservationRecord will still be used
but it's observation_status may be 15' out of date and the cadence
will run according to the potentially out-of-date status (i.e. a
PENDING observation that, if updated, would be WINDOW_EXPIRED, won't
be re-submitted, until the status is updated from the facility).